fix to build and run after update to angular v10
