+++ cms-server/package.json
@@ -14,9 +14,9 @@
   },
   "author": "danghung1202@gmail.com",
   "dependencies": {
-    "@angular-cms/api": "0.0.1",
+    "@angular-cms/api": "1.0.0",
     "express": "^4.17.1",
-    "tslib": "^1.11.1"
+    "tslib": "^2.0.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.1",
@@ -24,6 +24,6 @@
     "concurrently": "^3.5.1",
     "nodemon": "^1.19.2",
     "ts-node": "^8.4.1",
-    "typescript": "~3.5.3"
+    "typescript": "^3.9.7"
   }
 }
