+++ cms/cms-demo/src/app/app.module.ts
@@ -1,17 +1,18 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, Renderer2, RendererFactory2 } from '@angular/core';
-import { TransferHttpCacheModule } from '@nguniversal/common';
-
 import { AngularCms, AuthModule, PAGE_AFTER_INIT } from '@angular-cms/core';
-
-import * as contentTypes from './register-content-types';
+import { DOCUMENT } from '@angular/common';
+import { HttpClientModule } from '@angular/common/http';
+import { NgModule, RendererFactory2 } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { TransferHttpCacheModule } from '@nguniversal/common';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LayoutComponent } from './shared/layout/layout.component';
-import { PagesModule } from './pages/pages.module';
 import { BlocksModule } from './blocks/block.module';
-import { HttpClientModule } from '@angular/common/http';
-import { DOCUMENT } from '@angular/common';
+import { PagesModule } from './pages/pages.module';
+import * as contentTypes from './register-content-types';
+import { LayoutComponent } from './shared/layout/layout.component';
+
+
 
 AngularCms.registerContentTypes(contentTypes);
 
@@ -33,6 +34,7 @@ export function pageAfterViewInit(rendererFactory: RendererFactory2, document: D
 
 @NgModule({
     imports: [
+        BrowserAnimationsModule,
         BrowserModule.withServerTransition({ appId: 'serverApp' }),
         TransferHttpCacheModule,
         HttpClientModule,
