+++ cms/package.json
@@ -25,7 +25,7 @@
     "@angular/platform-browser-dynamic": "^10.2.4",
     "@angular/platform-server": "^10.2.4",
     "@angular/router": "^10.2.4",
-    "@fortawesome/angular-fontawesome": "^0.5.0",
+    "@fortawesome/angular-fontawesome": "^0.7.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.26",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
     "@nguniversal/common": "^9.0.2",
@@ -43,6 +43,8 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
+    "@angular-devkit/build-angular": "^0.1002.1",
+    "@angular-devkit/build-ng-packagr": "^0.1002.0",
     "@angular/cli": "^10.2.1",
     "@angular/compiler-cli": "^10.2.4",
     "@angular/language-service": "^10.2.4",
@@ -50,6 +52,7 @@
     "@types/jasmine": "^3.3.8",
     "@types/jasminewd2": "^2.0.3",
     "@types/node": "^12.11.1",
+    "@types/quill": "^1.3.10",
     "codelyzer": "^5.1.2",
     "concurrently": "^5.0.2",
     "install": "^0.13.0",
